more scripts + simplification
