+++ scripts/project/config.json
@@ -19,7 +19,7 @@
     {
       "documentation": {
         "command": "test-worship-harvest",
-        "description": "Test: harvest 1 dummy mandataris + 1 bedienaar for one eredienst-bestuur and verify publication",
+        "description": "Test: harvest 1 dummy mandataris + 1 bedienaar for one eredienst-bestuur (no publication)",
         "arguments": ["[bestuur-uri]"]
       },
       "environment": {
@@ -27,6 +27,20 @@
         "interactive": true,
         "join_networks": true,
         "script": "test-worship-harvest/run.sh"
+      },
+      "mounts": { "app": "/data/app/" }
+    },
+    {
+      "documentation": {
+        "command": "test-worship-harvest-and-publish",
+        "description": "Test: harvest 1 dummy mandataris + 1 bedienaar for one eredienst-bestuur and verify publication",
+        "arguments": ["[bestuur-uri]"]
+      },
+      "environment": {
+        "image": "node:20-bookworm",
+        "interactive": true,
+        "join_networks": true,
+        "script": "test-worship-harvest/run-and-publish.sh"
       },
       "mounts": { "app": "/data/app/" }
     }
