+++ scripts/project/test-worship-harvest/run-and-publish.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+export JOB_OPERATION="http://lblod.data.gift/id/jobs/concept/JobOperation/lblodHarvestWorshipAndPublish"
+exec node /script/main.js "$@"
